[Workflows][US_ME] Move resident record fields to state-specific metadata (Recidiviz/recidiviz-data#30587)
## Description of the change

The resident record has three fields (`portion_served_needed`,
`us_me_portion_needed_eligible_date`, and `sccp_eligibility_date`) that
are only used in Maine. This PR moves those fields into the `metadata`
field on the resident record. Additionally, this PR fixes a bug in those
fields that caused nulls to be parsed as the string "null".

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30546.
Related to Recidiviz/recidiviz-data#30584

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 920d29c9c1ccfff0575f5ed09a4f194d645a2ef8