[US_AZ] Add ADC_NUMBER to person_external_id (Recidiviz/recidiviz-data#30456)
## Description of the change

There is only one ID that seems to be meaningful to TTs in AZ, which is
the `ADC_NUMBER`. This ID isn't found on any tables except for one in
the ACIS data; since it is not particularly useful to us, I chose its
more widely-used counterpart (`PERSON_ID`) as a substitute. This is
making referencing people with TTs difficult, so this change adds the
ADC_NUMBER to `state_person_external_id` for reference.

I did that by updating every existing ingest view that creates an entity
that attaches to a `state_person` entity to also pull in that person's
`ADC_NUMBER`, so that the `ADC_NUMBER` can be included in their
`external_id` in the ingest mapping.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30411

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c28b2bfc4de285f5593fdbe42a3a69ccdb7670a1