update raw data table schema generation (Recidiviz/recidiviz-data#30588)
## Description of the change

motivation of this started in moving generating raw data table schemas
out of `DirectIngestRawFileImportManager`. in an effort to do so, this
PR:
- moves column description trimming logic to
`big_query_utils`/`constants`
- makes a new class `RawDataTableBigQuerySchemaBuilder` to centralize &
(hopefully) simply schema generation; it's interface is --
- `build_bq_schmea_for_config` -- just looks at the raw file config and
generates what we expect the big query schema to be
- `build_bq_schema_for_config_from_columns` -- allows you to pass in a
set of columns. not used right now, but will be used in cases where
`allow_incomplete_configs` is `True` (we dont know w/o looking at the
file itself what all the columns are)
- `build_bq_schmea_from_columns` -- to be deprecated, but a method that
mirrors
`DirectIngestRawFileImportManager.create_raw_table_schema_from_columns`
that will be removed once that class is removed

In working on this, I also noticed that the `allow_incomplete_configs`
flag, when uploading files w/ fields not present, is broken (see Recidiviz/recidiviz-data#30562)
in `import_raw_files_to_bq_sandbox` due to how we fetch column
descriptions (and also how we create tables schemas). Since it was only
a two line change to fix that issue, I included that in this PR as well.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30562
Prework for Recidiviz/recidiviz-data#30168

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b65c64870759978527c4b5d5ffff9b6880a44c63