[Justice Counts] Add a script to test parity between the Metric Settings table and the Agency Datapoints. (Recidiviz/recidiviz-data#29919)
## Description of the change

Add a script to test parity between the Metric Settings table and the
Agency Datapoints.

Zenhub ticket Recidiviz/recidiviz-data#28420

Successful invocation of this script means that the metric interfaces
present as agency
datapoints exactly match the metric interfaces present in the metric
settings table.
We will use this script to routinely check for parity between the two
tables, during the
period that we are double-writing to both tables.

This job
- Reads all agency metric interfaces from the datapoints and the metric
settings tables.
- Compares the metric interfaces using DeepDiff (after converting them
to json using to_json).
- Raises any errors if the metric interfaces do not match.

Other changes in this PR:
* Renaming the other metric settings jobs to be prefixed with
"metric_settings_" so that they are easier to identify.

## Testing
I ran this parity script in local and staging environments, after
calling the overwrite_metric_settings script on those same tables.

GitOrigin-RevId: af0007983870cdc46daed9325c1d29e3fb44d023