insert json rows in one statement when importing to cloud sql (Recidiviz/recidiviz-data#30747)
## Description of the change

The `import/insights/US_TN/supervision_client_events.json` endpoint is
consistently timing out. As per
https://recidiviz.slack.com/archives/G01P6S547T4/p1718637367715969,
attempt to speed it up by passing all the entities at once, which does a
single INSERT call instead of one per entity.

Tested it with the load_local_db script. The timing was pretty
inconsistent between runs, but the new version seems to be a 2-3x
speedup compared with the old so hopefully this prevents it from timing
out going forward.

Something also interesting: the TN supervision_client_events.json table
took only ~50 seconds to import locally using the old version of the
script . It's weird that the cloud run version is more than 10x slower!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 77e4d50f39522c2fe42eab5b11d4f7bea98156c8