[US_ID][JII] Convert analytics script into notebook (Recidiviz/recidiviz-data#30529)
## Description of the change

Convert/copy the existing
[id_lsu_message_analytics.py](https://github.com/Recidiviz/recidiviz-data/blob/main/recidiviz/case_triage/jii/id_lsu_message_analytics.py)
script into a jupyter notebook. This will allow for easier/faster
exploratory queries and one-off analysis after future launches. The only
difference between this notebook and the existing script is some added
cells with example queries / a place to play around with querying the
database.

#### How I Tested

I was able to run the notebook and successfully connect to/query the
database

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@Michelles-MBP.attlocal.net>
GitOrigin-RevId: da93c02f5e6256fc3dfc0568ac45490408a851d2