[Validators] Add attr_validator that checks that items in a list have the correct type (Recidiviz/recidiviz-data#30776)
## Description of the change

Adds a new `is_list_of()` validator that can be used to make sure that
all items in a list on a list field have the expected type. I am going
to use this in my new implementation of normalized entities (list fields
currently have no validation for list item types).

The new validator is implemented as a class under the hood so I can test
in a future PR that all list fields have this particular validator set.
I also updated the `is_opt` validator to use a class under the hood so I
can check if a validator allows for optionals in a test as well.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#30075

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f3f52f92192e597b1b15eced0860aa4625b07478